fix: rename Sidebar file
